Start of refactoring / overhaul
Current status: not fully working
                model building and testing in progress

.plan:

Better structure:
Instead of directly manipulating a big data dictionary,
data is now represented in a class based structure in memory.

Better models:
The addon class better models the data found in the addon description.
The parser is moved to the addon class and extracts all official
data fields, even if the tool does not use them by now.
Just in case I might need them later --
or in case you want to use the parser for own projects.

The dependency class handles the dependency description,
keeps track of minimum and maximum versions and will contain all
version checking and dependency merging logic

The DependencyMatrix class takes the list of scanned addons and
tries to deep scan their dependencies on multiple levels
After scanning all dependencies, it tries to reduces the information
by merging duplicate findings. When merging, the more restrictive
version requirement is kept and the less restrictive is discarded.

After building the DependencyMatrix, it can be used to report
- unresolved mandatory dependencies
- unresolved optional dependencies
- libraries that are not used by any other addon